Native staking tests
