+++ anchor/fixtures/accounts/J2nUHEAgZFRyuJbFjdqPrAa9gyWDuc7hErtDQHPhsYRp.json
@@ -1,0 +1,14 @@
+{
+  "pubkey": "J2nUHEAgZFRyuJbFjdqPrAa9gyWDuc7hErtDQHPhsYRp",
+  "account": {
+    "lamports": 3248590146603,
+    "data": [
+      "AgAAAF0u2POwTKnQAengBBUCFU30jYkx82QZfG+TzZN4VySTP+pdoho3k8fzwkvupc0lZIKppE6fnvy7AORiK2C5/GIJHwAAAAAAAAAAYUSpEAAAAAAfAAAAAGJEqRAAAAAAHgAAAABjRKkQAAAAAB0AAAAAZESpEAAAAAAcAAAAAGVEqRAAAAAAGwAAAABmRKkQAAAAABoAAAAAZ0SpEAAAAAAZAAAAAGhEqRAAAAAAGAAAAABpRKkQAAAAABcAAAAAakSpEAAAAAAWAAAAAGtEqRAAAAAAFQAAAABsRKkQAAAAABQAAAAAbUSpEAAAAAATAAAAAG5EqRAAAAAAEgAAAABvRKkQAAAAABEAAAAAcESpEAAAAAAQAAAAAHFEqRAAAAAADwAAAAByRKkQAAAAAA4AAAAAc0SpEAAAAAANAAAAAHREqRAAAAAADAAAAAB1RKkQAAAAAAsAAAAAdkSpEAAAAAAKAAAAAHdEqRAAAAAACQAAAAB4RKkQAAAAAAgAAAAAeUSpEAAAAAAHAAAAAHpEqRAAAAAABgAAAAB7RKkQAAAAAAUAAAAAfESpEAAAAAAEAAAAAH1EqRAAAAAAAwAAAAB+RKkQAAAAAAIAAAAAf0SpEAAAAAABAAAAAWBEqRAAAAAAAQAAAAAAAACHAgAAAAAAAF0u2POwTKnQAengBBUCFU30jYkx82QZfG+TzZN4VySTAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHwAAAAAAAAABQAAAAAAAAABIAgAAAAAAAAwegQQAAAAAbuF6BAAAAABJAgAAAAAAAKo1hwQAAAAADB6BBAAAAABKAgAAAAAAAOYxjQQAAAAAqjWHBAAAAABLAgAAAAAAAEwZkwQAAAAA5jGNBAAAAABMAgAAAAAAAHzRmAQAAAAATBmTBAAAAABNAgAAAAAAANuPngQAAAAAfNGYBAAAAABOAgAAAAAAANJwpAQAAAAA24+eBAAAAABPAgAAAAAAAIl/qgQAAAAA0nCkBAAAAABQAgAAAAAAAI17sAQAAAAAiX+qBAAAAABRAgAAAAAAAF2XtgQAAAAAjXuwBAAAAABSAgAAAAAAAOGevAQAAAAAXZe2BAAAAABTAgAAAAAAABOOwgQAAAAA4Z68BAAAAABUAgAAAAAAAJ+QyAQAAAAAE47CBAAAAABVAgAAAAAAADuDzgQAAAAAn5DIBAAAAABWAgAAAAAAAEgC1AQAAAAAO4POBAAAAABXAgAAAAAAAHRL2QQAAAAASALUBAAAAABYAgAAAAAAAHJu3gQAAAAAdEvZBAAAAABZAgAAAAAAAKPX4wQAAAAAcm7eBAAAAABaAgAAAAAAAO6T6QQAAAAAo9fjBAAAAABbAgAAAAAAAExd7wQAAAAA7pPpBAAAAABcAgAAAAAAALFB9QQAAAAATF3vBAAAAABdAgAAAAAAAPUX+wQAAAAAsUH1BAAAAABeAgAAAAAAAI7mAAUAAAAA9Rf7BAAAAABfAgAAAAAAAJttBgUAAAAAjuYABQAAAABgAgAAAAAAAN4NDAUAAAAAm20GBQAAAABhAgAAAAAAANHqEQUAAAAA3g0MBQAAAABiAgAAAAAAAELNFwUAAAAA0eoRBQAAAABjAgAAAAAAAJm6HQUAAAAAQs0XBQAAAABkAgAAAAAAAEzCIwUAAAAAmbodBQAAAABlAgAAAAAAALfcKQUAAAAATMIjBQAAAABmAgAAAAAAAN4PMAUAAAAAt9wpBQAAAABnAgAAAAAAAGI7NgUAAAAA3g8wBQAAAABoAgAAAAAAALxhPAUAAAAAYjs2BQAAAABpAgAAAAAAAM+CQgUAAAAAvGE8BQAAAABqAgAAAAAAAP+aSAUAAAAAz4JCBQAAAABrAgAAAAAAAN+WTgUAAAAA/5pIBQAAAABsAgAAAAAAACaOVAUAAAAA35ZOBQAAAABtAgAAAAAAAMxkWgUAAAAAJo5UBQAAAABuAgAAAAAAALUwYAUAAAAAzGRaBQAAAABvAgAAAAAAAIUEZgUAAAAAtTBgBQAAAABwAgAAAAAAAIjoawUAAAAAhQRmBQAAAABxAgAAAAAAAEvNcQUAAAAAiOhrBQAAAAByAgAAAAAAAAT4dwUAAAAAS81xBQAAAABzAgAAAAAAANbjfQUAAAAABPh3BQAAAAB0AgAAAAAAAH7rgwUAAAAA1uN9BQAAAAB1AgAAAAAAAFEWigUAAAAAfuuDBQAAAAB2AgAAAAAAAHslkAUAAAAAURaKBQAAAAB3AgAAAAAAAGpOlgUAAAAAeyWQBQAAAAB4AgAAAAAAAERsnAUAAAAAak6WBQAAAAB5AgAAAAAAAPGHogUAAAAARGycBQAAAAB6AgAAAAAAAOaHqAUAAAAA8YeiBQAAAAB7AgAAAAAAAKWJrgUAAAAA5oeoBQAAAAB8AgAAAAAAAD+MtAUAAAAApYmuBQAAAAB9AgAAAAAAAOaRugUAAAAAP4y0BQAAAAB+AgAAAAAAAOKewAUAAAAA5pG6BQAAAAB/AgAAAAAAAHylxgUAAAAA4p7ABQAAAACAAgAAAAAAAPWlzAUAAAAAfKXGBQAAAACBAgAAAAAAAL6x0gUAAAAA9aXMBQAAAACCAgAAAAAAAM7T2AUAAAAAvrHSBQAAAACDAgAAAAAAALvq3gUAAAAAztPYBQAAAACEAgAAAAAAANPp5AUAAAAAu+reBQAAAACFAgAAAAAAAL8U6wUAAAAA0+nkBQAAAACGAgAAAAAAAPst8QUAAAAAvxTrBQAAAACHAgAAAAAAAJ+F8QUAAAAA+y3xBQAAAAB/RKkQAAAAAODVoWYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+      "base64"
+    ],
+    "owner": "Vote111111111111111111111111111111111111111",
+    "executable": false,
+    "rentEpoch": 18446744073709551615,
+    "space": 3762
+  }
+}
