+++ anchor/programs/glam/src/lib.rs
@@ -120,14 +120,14 @@ pub mod glam {
 
     // Stake pool
     pub fn stake_pool_deposit(ctx: Context<StakePoolDeposit>, lamports: u64) -> Result<()> {
-        staking::stake_pool_deposit(ctx, lamports)
+        stake_pool::stake_pool_deposit(ctx, lamports)
     }
 
     pub fn stake_pool_withdraw_sol(
         ctx: Context<StakePoolWithdrawSol>,
         lamports: u64,
     ) -> Result<()> {
-        staking::stake_pool_withdraw_sol(ctx, lamports)
+        stake_pool::stake_pool_withdraw_sol(ctx, lamports)
     }
 
     pub fn stake_pool_withdraw_stake(
@@ -136,7 +136,7 @@ pub mod glam {
         stake_account_id: String,
         stake_account_bump: u8,
     ) -> Result<()> {
-        staking::stake_pool_withdraw_stake(
+        stake_pool::stake_pool_withdraw_stake(
             ctx,
             pool_token_amount,
             stake_account_bump,
@@ -144,25 +144,26 @@ pub mod glam {
         )
     }
 
-    pub fn native_stake<'info>(
-        ctx: Context<'_, '_, '_, 'info, NativeStake<'info>>,
+    // Native staking
+    pub fn native_stake_deposit<'info>(
+        ctx: Context<'_, '_, '_, 'info, NativeStakeDeposit<'info>>,
         lamports: u64,
         stake_account_id: String,
         stake_account_bump: u8,
     ) -> Result<()> {
-        staking::native_stake(ctx, lamports, stake_account_id, stake_account_bump)
+        stake::native_stake_deposit(ctx, lamports, stake_account_id, stake_account_bump)
     }
 
     pub fn deactivate_stake_accounts<'info>(
         ctx: Context<'_, '_, '_, 'info, DeactivateStakeAccounts<'info>>,
     ) -> Result<()> {
-        staking::deactivate_stake_accounts(ctx)
+        stake::deactivate_stake_accounts(ctx)
     }
 
     pub fn withdraw_from_stake_accounts<'info>(
         ctx: Context<'_, '_, '_, 'info, WithdrawFromStakeAccounts<'info>>,
     ) -> Result<()> {
-        staking::withdraw_from_stake_accounts(ctx)
+        stake::withdraw_from_stake_accounts(ctx)
     }
 
     // Jupiter
